// src/app/models/voucher.model.ts
export interface Voucher{
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  tel: string;
  numOfChildren: number;
  numOfAdults: number;
  numOfBedrooms: number;
  numOfNights: number;
  country: string;
  postcode: number;
  city: string;
  address: string;
  isPaid: boolean;
}